# server/common/swagger/Api.yaml
swagger: '2.0'
info:
  version: 1.10.1
  title: nxplorerjs-microservice
  description: Express Microservice Starter
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and retourned as answer of a call
      to the route /login giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

          Bearer xxxxxx.yyyyyyy.zzzzzz
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  PublishedProductsBody:
    type: object
    title: published products
    required:
      - name
      - data
    properties:
      name:
        type: string
        description: Product List Name
      data:
        type: array
        items:
          $ref: '#/definitions/publish'
  publish:
    type: object
    title: published products
    required:
      - id
      - url
      - buildId
      - buildStatus
      - scrapeDate   
    properties:
      id:
        type: string
        description: publish name
      url:
        type: string
        description: build url
      buildStatus:
        type: string
        description: build status
        enum: [inprogress, successful, failure, notstarted]
      buildId:
        type: string
        description: build id
      scrapeDate:
        type: string
        description: date scraped
        format: date-time
      publishStartDate:
        type: string
        description: publish start date 
        format: date-time
      publishEndDate:
        type: string
        description: publish end date
        format: date-time
  ProductsBody:
    type: object
    title: products
    required:
      - name
      - data
    properties:
      name:
        type: string
        description: Product List Name
      data:
        type: array
        items:
          $ref: '#/definitions/product'
      country:
        type: string
        description: campaign country
      theme:
        type: object
        title: Page Theme
        properties:
          page:
            type: object
            title: page theme
            properties:
              backgroundImage: 
                type: string
                description: background Image
              background:
                type: string
                description: background color
      visibility:
        type: object
        title: page visibility details
        properties:
          startDate:
            type: string
            description: page start date
            format: date-time
          endDate:
            type: string
            description: page end date
            format: date-time
  product:
    type: object
    title: scraped product
    required:
      - title
      - salePrice
      - brand
      - availability
      - vnv
      - brandUrl
      - image
    properties:
      title:
        type: string
        description: product name
      salePrice:
        type: string
        description: sale price        
      brand:
        type: string
        description: brand name
      availability:
        type: string
        description: availability
      vnv:
        type: string
        description: veg / non veg
      brandUrl:
        type: string
        description: brand url
      image:
        type: string
        description: base64 Image string
      features:
        type: array
        items:
          $ref: '#/definitions/productFeature'
      images:
        type: array
        items:
          $ref: '#/definitions/productImage'
  productFeature:
    type: object
    title: feature line
    required:
      - feature
    properties:
      feature:
        type: string
        description: feature line
  productImage:
    type: object
    title: image details
    required:
      - url
    properties:
      url:
        type: string
        description: image url                     
...
paths:
...
  /scraper/scrape:
    get:
      tags:
        - Scrape URL
      description: Returns the product scraped from Amazon
      parameters:
        - name: name
          in: query
          required: true
          description: Name of the List
          type: string
        - name: headless
          in: query
          required: false
          type: boolean
        - name: url
          in: query
          required: true
          description: URL to scrape for example (https://www.amazon.in/dp/B074ZJY1JD)
          type: string
        - name: save
          in: query
          required: true
          description: Persist the list
          type: boolean
      responses:
        200:
          description: Gets product data that is scraped from Amazon
  /scraper/scrapeList:
    get:
      tags:
        - Scrape Amazon ASIN LIST
      description: Returns the list products scraped from Amazon
      parameters:
        - name: name
          in: query
          required: true
          description: Name of the List
          type: string
        - name: country
          in: query
          required: false
          type: string
        - name: marketplace
          in: query
          required: false
          type: string
        - name: baseUrl
          in: query
          required: false
          type: string
        - name: asinList
          in: query
          required: true
          description: Comma separated ASIN List to scrape for example "B074ZJY1JD,B01GCEF7R8"
          type: string
        - name: startDate
          required: false
          in: query
          type: string
          description: publish start date
          format: date-time
        - name: endDate
          required: false
          in: query
          type: string
          description: publish end date
          format: date-time
        - name: save
          in: query
          required: true
          description: Persist the list
          type: boolean
        - name: withURL
          in: query
          required: false
          description: If full URL is provided
          type: boolean
      responses:
        200:
          description: Gets product data that is scraped from Amazon
  /scraper/products:
    get:
      tags:
        - Get all products
      responses:
        200:
          description: Return the products with the specified name
        404:
          description: Products not available   
    post:
      tags:
        - Save products
      description: Save Scrapped Products Data JSON
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: product JSON
          required: true
          schema:
            $ref: '#/definitions/ProductsBody'
      responses:
        201:
          description: Products List posted
        403:
          description: Forbidden Request
  /scraper/publish:
    get:
      tags:
        - Get all published products
      responses:
        200:
          description: Return the products with the specified name
        404:
          description: Products not available   
    post:
      tags:
        - Save publish details
      description: Save Published Products Data JSON
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: product JSON
          required: true
          schema:
            $ref: '#/definitions/PublishedProductsBody'
      responses:
        201:
          description: Published Products List posted
        403:
          description: Forbidden Request
  /scraper/publish/{name}:
    get:
      tags:
        - Get publish details
      parameters:
        - name: name
          in: path
          required: true
          description: The name of the published site
          type: string
      responses:
        200:
          description: Return the published site with the specified name
        404:
          description: Publish site not available
    delete:
      description: deletes a single campaign publish record based on the name supplied
      operationId: deleteCampaignPublish
      parameters:
        - name: name
          in: path
          description: Name of campaign publish record to delete
          required: true
          type: string
      responses:
        204:
          description: campaign publish record deleted
        404:
          description: campaign publish record not available      
  /scraper/products/{name}:
    get:
      tags:
        - Get products
      parameters:
        - name: name
          in: path
          required: true
          description: The name of the entity to retrieve
          type: string
      responses:
        200:
          description: Return the products with the specified name
        404:
          description: Products not available
    delete:
      description: deletes a single can campaign based on the name supplied
      operationId: deleteCampaignProducts
      parameters:
        - name: name
          in: path
          description: Name of campaign to delete
          required: true
          type: string
      responses:
        204:
          description: campaign deleted
        404:
          description: Campaign not available           